sales_analyst/requirements.txt
langchain-core~=1.2.7
python-dotenv~=1.2.1
langchain-openai~=1.1.7
langchain~=1.2.7
openai~=2.15.0
numpy~=2.4.1
matplotlib~=3.10.8
autogen[all]